cxx-qt-lib: seal QFlagRepr trait
