+++ crates/cxx-qt-lib/src/core/qflags/qflag.rs
@@ -3,58 +3,9 @@
 //
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-use std::fmt::Debug;
-use std::hash::Hash;
-use std::ops::{BitAnd, BitAndAssign, BitOr, BitOrAssign, BitXor, BitXorAssign, Not};
+use super::QFlagRepr;
 
 use cxx::ExternType;
-
-pub trait QFlagRepr: Sized {
-    /// Qt chooses the integer representation for a `QFlags<T>` as follows:
-    ///
-    /// - If `T` is signed, use a signed integer. Otherwise, use an unsigned integer.
-    /// - If `T` is 32 bits or less, use a 32-bit integer.
-    /// - If `T` is 64 bits and the Qt version is at least 6.9, use a 64-bit integer.
-    type Int: From<Self>
-        + Copy
-        + Debug
-        + Default
-        + Eq
-        + Ord
-        + Hash
-        + BitAnd<Output = Self::Int>
-        + BitAndAssign
-        + BitOr<Output = Self::Int>
-        + BitOrAssign
-        + BitXor<Output = Self::Int>
-        + BitXorAssign
-        + Not<Output = Self::Int>
-        + ExternType<Kind = cxx::kind::Trivial>;
-
-    const ZERO: Self::Int;
-}
-
-macro_rules! impl_repr {
-    ($t:ty, $i:ty) => {
-        impl QFlagRepr for $t {
-            type Int = $i;
-
-            const ZERO: Self::Int = 0;
-        }
-    };
-}
-
-impl_repr!(i8, i32);
-impl_repr!(i16, i32);
-impl_repr!(i32, i32);
-impl_repr!(u8, u32);
-impl_repr!(u16, u32);
-impl_repr!(u32, u32);
-
-#[cfg(cxxqt_qt_version_at_least_6_9)]
-impl_repr!(i64, i64);
-#[cfg(cxxqt_qt_version_at_least_6_9)]
-impl_repr!(u64, u64);
 
 /// # Safety
 ///
@@ -90,17 +41,13 @@ pub unsafe trait QFlag: Sized {
 }
 
 /// Internal utility trait for converting `T` in a `QFlag<T>` to the corresponding integer type.
-pub trait QFlagExt: QFlag {
-    type Int;
-
-    fn to_int(self) -> Self::Int;
+pub(super) trait QFlagExt: QFlag {
+    fn to_int(self) -> <Self::Repr as QFlagRepr>::Int;
 }
 
 impl<T: QFlag> QFlagExt for T {
-    type Int = <<T as QFlag>::Repr as QFlagRepr>::Int;
-
     #[inline(always)]
-    fn to_int(self) -> Self::Int {
+    fn to_int(self) -> <Self::Repr as QFlagRepr>::Int {
         self.to_repr().into()
     }
 }
